Move the current options into a menu so that its easier to add the option to view deliveries
