+++ shared/i18n/locales/en_US/translation.json
@@ -321,6 +321,8 @@
   "Revoke invite": "Revoke invite",
   "Activate account": "Activate account",
   "Suspend account": "Suspend account",
+  "Delete webhook": "Delete webhook",
+  "Edit webhook": "Edit webhook",
   "API token created": "API token created",
   "Name your token something that will help you to remember it's use in the future, for example \"local development\", \"production\", or \"continuous integration\".": "Name your token something that will help you to remember it's use in the future, for example \"local development\", \"production\", or \"continuous integration\".",
   "The document archive is empty at the moment.": "The document archive is empty at the moment.",
@@ -601,10 +603,8 @@
   "URL": "URL",
   "All events": "All events",
   "All {{ groupName }} events": "All {{ groupName }} events",
-  "Delete webhook": "Delete webhook",
   "Disabled": "Disabled",
   "Subscribed events": "Subscribed events",
-  "Edit webhook": "Edit webhook",
   "Webhook created": "Webhook created",
   "Logo updated": "Logo updated",
   "Unable to upload new logo": "Unable to upload new logo",
